refactor: split game state reducers by responsibility
